GPII-1542: Remove custom Write Defaults failure handling
It's not safe to add and remove failure handlers on a per-request basis
in the way that I was doing. The only real option would be to override
the failure handler permanently. However, with the current Fluid failure
system, that doesn't really buy us anything as it is not possible to
differentiate between failure types.